streamlit/cgm_app/requirements.txt
streamlit
pandas
pillow
streamlit-aggrid==0.2.2-2
click==8
plotly